+++ actix-http/src/h1/encoder.rs
@@ -76,6 +76,14 @@ pub(crate) trait MessageType: Sized {
                     skip_len = true;
                     length = BodySize::None
                 }
+
+                StatusCode::NOT_MODIFIED => {
+                    // don't skip content-length header for not modified responses
+                    // see https://datatracker.ietf.org/doc/html/rfc7232#section-4.1
+                    skip_len = false;
+                    length = BodySize::None;
+                }
+
                 _ => {}
             }
         }